Move via_url to view-specific helpers
