+++ lms/views/helpers/__init__.py
@@ -1,4 +1,5 @@
 from lms.views.helpers._canvas_files import canvas_files_available
+from lms.views.helpers._via import via_url
 
 
-__all__ = ("canvas_files_available",)
+__all__ = ["canvas_files_available", "via_url"]
